fix the bug of oneway and deserialization
