+++ codec-common/src/main/java/io/esastack/codec/common/utils/DefaultValueUtils.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2021 OPPO ESA Stack Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.esastack.codec.common.utils;
+
+public class DefaultValueUtils {
+    public static Object getUnboxedPrimitiveDefaultValue(final Class<?> returnType) {
+        if (returnType != null && returnType.isPrimitive()) {
+            if (returnType == byte.class) {
+                return (byte) 0;
+            }
+            if (returnType == short.class) {
+                return (short) 0;
+            }
+            if (returnType == char.class) {
+                return Character.MIN_VALUE;
+            }
+            if (returnType == int.class) {
+                return 0;
+            }
+            if (returnType == long.class) {
+                return 0L;
+            }
+            if (returnType == boolean.class) {
+                return false;
+            }
+            if (returnType == float.class) {
+                return 0.0f;
+            }
+            if (returnType == double.class) {
+                return 0.0d;
+            }
+        }
+        return null;
+    }
+}
